+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Added / Changed / Deprecated / Removed / Fixed / Security
 
 - The `getDistribution()` method in the TDistribution is now implemented for the general case. Previously it was only implemented for a few distinct degrees of freedom.
 
+### Changed
+
+- Generic declarations interfaces and abstract classes (everywhere) that used to be `<N extends Number>` are now `<N extends Comparable<N>>`.
+
 ## [47.3.1] – 2019-09-29
 
 ### Added